requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
lxml
python-dotenv
